remove duplicate odld propagator from tests
